Fix some commpiler warnings.
* bodyio.h (decode_line): return size_t, not int, for a length.
(encode_line): take a size_t parameter, not int, for a length.
* encoding.cc (decode_line): Use a size_t value for offsets in the
buffer.
* bodyio.cc (body_insert_text): Add casts where we know
conversions are safe to avoid compiler warnings.
(copy_data): Mark as a static function.
* mylist.h (mylist): Remove empty constructor body definition.

+++ bodyio.h
@@ -43,8 +43,8 @@ bool body_insert(bool *binary,
 
 
 // Encoding functions...
-void encode_line(const char in[], char out[], int len);
-int decode_line(const char in[], char out[]);
+void encode_line(const char in[], char out[], size_t len);
+size_t decode_line(const char in[], char out[]);
 int encode_stream(FILE *fin, FILE *fout); //encodes whole file.
